Move create class
